services/api-gateway/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# HTTP client for service communication
httpx==0.25.2

# Authentication
pyjwt==2.8.0

# Monitoring
prometheus-client==0.19.0

# Database drivers for health checks
redis==5.0.1
asyncpg==0.29.0

# Utilities
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0